pyproject.toml
[tool.poetry]
name = "chefrag"
version = "0.1.0"
description = ""
authors = ["epishchik <evgeniipishchik@mail.ru>"]
license = "apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.67.1"
beautifulsoup4 = "^4.13.3"
requests = "^2.32.3"
pre-commit = "^4.1.0"
pandas = "^2.2.3"
pyarrow = "^19.0.0"
python-dotenv = "^1.0.1"
datasets = "^3.3.1"
openai = "^1.63.2"
ollama = "^0.4.7"
qdrant-client = "^1.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
